pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "project-forge"
version = "0.1.0"
description = "A Python tool for creating and managing project structures"
readme = "README.md"
authors = [{ name = "thatch", email = "gnarzilla@deadlight.boo" }]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]
dependencies = [
    "click >= 8.0.0",
    "rich >= 13.0.0",
    "jinja2 >= 3.0.0",
    "pydantic >= 2.0.0",
    "PyYAML >= 6.0.0",
    "toml >= 0.10.0",
    "packaging >= 23.0",
    "black >= 23.0.0",   # Add these
    "isort >= 5.0.0"     # Add these
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pytest >= 7.0.0",
    "pytest-cov >= 4.0.0",
    "black >= 23.0.0",
    "isort >= 5.0.0",
    "mypy >= 1.0.0"
]

[project.urls]
Homepage = "https://github.com/thatch/project-forge"
Documentation = "https://project-forge.readthedocs.io/"
Repository = "https://github.com/thatch/project-forge.git"

[project.scripts]
project-forge = "project_forge.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/project_forge"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.project-forge]
package_excludes = [
    "__pycache__",
    "templates",
    "examples",
    "tests",
    "archive"
]
validation_dirs = [
    "src/project_forge",
    "src/project_forge/cli",
    "src/project_forge/cli/commands",
    "src/project_forge/utils",
    "src/project_forge/core"
]
